LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/prefers-reduced-data-expected.txt


FAIL Should be parseable in a CSS stylesheet: '(prefers-reduced-data)' assert_true: expected true got false
FAIL Should be parseable in a CSS stylesheet: '(prefers-reduced-data: no-preference)' assert_true: expected true got false
FAIL Should be parseable in a CSS stylesheet: '(prefers-reduced-data: reduce)' assert_true: expected true got false
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: 0)' 
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: none)' 
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: 10px)' 
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: no-preference reduce)' 
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: reduced)' 
PASS Should not be parseable in a CSS stylesheet: '(prefers-reduced-data: no-preference/reduce)' 
FAIL Should be parseable in JS: '(prefers-reduced-data)' assert_true: expected true got false
FAIL Should be parseable in JS: '(prefers-reduced-data: no-preference)' assert_true: expected true got false
FAIL Should be parseable in JS: '(prefers-reduced-data: reduce)' assert_true: expected true got false
PASS Should not be parseable in JS: '(prefers-reduced-data: 0)' 
PASS Should not be parseable in JS: '(prefers-reduced-data: none)' 
PASS Should not be parseable in JS: '(prefers-reduced-data: 10px)' 
PASS Should not be parseable in JS: '(prefers-reduced-data: no-preference reduce)' 
PASS Should not be parseable in JS: '(prefers-reduced-data: reduced)' 
PASS Should not be parseable in JS: '(prefers-reduced-data: no-preference/reduce)' 
FAIL Check that no-preference evaluates to false in the boolean context assert_equals: expected true but got false
PASS Check that invalid evaluates to false 
